requirements.txt
numpy==1.26.4
torch==2.3.1+cpu
torchvision==0.18.1
pillow==10.4.0
matplotlib==3.2.2
voila==0.2.4
